handle_get() to truncate data on last "\nEND\n"
Searching the buffer string for "END" causes problems if the data inside contains the substring "END". For example, if I retrieve the string "Please do not EXTEND beyond the edge" from my memcached server, the "END" in EXTEND will signal the end of the data, and only "Please do Not E" will be returned. This is fixed by looking for ``crlf + "END" + crlf``. On the off chance that the data contains that string, it is advisable to search for the last occurrence of it, instead of the first.